\name{zelig2exp}
\alias{zelig2exp}
\title{Interface between the Zelig Model exp and...}
\usage{zelig2exp(formula, ..., robust=FALSE, cluster, data)}
\description{Interface between the Zelig Model exp and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
