\name{zelig2lognorm}
\alias{zelig2lognorm}
\title{Interface between the Zelig Model lognorm and...}
\usage{zelig2lognorm(formula, ..., data)}
\description{Interface between the Zelig Model lognorm and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
