\name{zelig2relogit}
\alias{zelig2relogit}
\title{Interface between the Zelig Model relogit and...}
\usage{zelig2relogit(formula, ..., data)}
\description{Interface between the Zelig Model relogit and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
