\name{zelig2rq}
\alias{zelig2rq}
\title{Interface between the Zelig Model rq and...}
\usage{zelig2rq(formula, ..., weights, tau=0.5, data)}
\description{Interface between the Zelig Model rq and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
