\name{zelig2tobit}
\alias{zelig2tobit}
\title{Interface between the Zelig Model tobit and...}
\usage{zelig2tobit(formula, ..., below=0, above=Inf, robust=FALSE, cluster,
    data)}
\description{Interface between the Zelig Model tobit and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
