\name{ZeligCommon-package}
\title{ZeligCommon...}
\description{ZeligCommon}
\details{\tabular{ll}{
Package: \tab ZeligCommon\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligCommon-package}
\alias{ZeligCommon}
\docType{package}
\author{Matt Owen}
\keyword{package}
